requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
dominate==2.5.1
Flask==1.1.1
Flask-Bootstrap4==4.0.2
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
requests==2.23.0
SQLAlchemy==1.3.15
urllib3==1.26.5
visitor==0.1.3
Werkzeug==1.0.0
WTForms==2.2.1